Add nonlinear traveling capabilities
Commands "advance" and "backtrack" were replaced by a single command "travel" which encourages a more nonlinear traveling method. This new command prints to you a list of locations to choose from, where you could travel to.

The field for Player that stores the current location was an index to a list of Location objects. This kind of scheme encourages very linear progression throughout the game. It was changed to use a pointer instead, with getter/setter methods, which will better support nonlinear progression.

A Location should keep track of where you can go from where you are right now. This gives off a kind of multiply-linked list behavior, letting you more freely create a nonlinear map to travel around in. You would also be able to have one-way entrances to other locations if that ever becomes necessary.