Add terminal-style input
Everytime before you enter in a command, the application will print the following line: "<name>:/world/<location>$ ". It looks pretty similar to a terminal-style input, except without colors obviously. To make it easier to distinguish from the rest of the text, a newline was added after each command was executed. Some of the string literals in main.cpp and game.cpp had to be slightly modified to support this change.